+++ src/Compilers/Core/Portable/SourceGeneration/Nodes/ISyntaxInputNode.cs
@@ -1,0 +1,20 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Collections.Immutable;
+
+namespace Microsoft.CodeAnalysis
+{
+    internal interface ISyntaxInputNode
+    {
+        ISyntaxInputBuilder GetBuilder(DriverStateTable table);
+    }
+
+    internal interface ISyntaxInputBuilder
+    {
+        void VisitTree(SyntaxNode root, EntryState state, SemanticModel? model);
+
+        void SaveStateAndFree(ImmutableDictionary<object, IStateTable>.Builder tables);
+    }
+}
